vp9: Fix integer overflow in ratectrl.c
Issue introduced here:
https://chromium-review.googlesource.com/c/webm/libvpx/+/6094823

Can be triggered with:
https://chromium-review.googlesource.com/c/webm/libvpx/+/6112963/2

Bug: 376707227

+++ vp9/encoder/vp9_ratectrl.c
@@ -1310,11 +1310,11 @@ static int rc_pick_q_and_bounds_one_pass_vbr(const VP9_COMP *cpi,
     // compensate for the pre-encoded buffer update (in
     // vp9_rc_get_one_pass_vbr_params).
     const int qp_thresh = 32;
-    const int bitrate_err =
-        (int)(cpi->framerate *
-              (rc->buffer_level - rc->starting_buffer_level -
-               rc->avg_frame_bandwidth) /
-              (cm->current_video_frame + 1));
+    const int64_t bitrate_err =
+        (int64_t)(cpi->framerate *
+                  (rc->buffer_level - rc->starting_buffer_level -
+                   rc->avg_frame_bandwidth) /
+                  (cm->current_video_frame + 1));
     // Threshold may be tuned, but for now condition this on low QP.
     if (cpi->oxcf.lag_in_frames == 0 && bitrate_err / 1000 < -10 &&
         (q < qp_thresh || *top_index < qp_thresh)) {
